+++ sendbird_platform_sdk/model/list_banned_channels_response_banned_channels_inner.py
@@ -88,10 +88,10 @@ class ListBannedChannelsResponseBannedChannelsInner(ModelNormal):
         """
         lazy_import()
         return {
-            'start_at': (float,),  # noqa: E501
+            'start_at': (int,),  # noqa: E501
             'description': (str,),  # noqa: E501
             'channel': (SendBirdChannelResponse,),  # noqa: E501
-            'end_at': (float,),  # noqa: E501
+            'end_at': (int,),  # noqa: E501
         }
 
     @cached_property
@@ -147,10 +147,10 @@ class ListBannedChannelsResponseBannedChannelsInner(ModelNormal):
                                 Animal class but this time we won't travel
                                 through its discriminator because we passed in
                                 _visited_composed_classes = (Animal,)
-            start_at (float): [optional]  # noqa: E501
+            start_at (int): [optional]  # noqa: E501
             description (str): [optional]  # noqa: E501
             channel (SendBirdChannelResponse): [optional]  # noqa: E501
-            end_at (float): [optional]  # noqa: E501
+            end_at (int): [optional]  # noqa: E501
         """
 
         _check_type = kwargs.pop('_check_type', True)
@@ -236,10 +236,10 @@ class ListBannedChannelsResponseBannedChannelsInner(ModelNormal):
                                 Animal class but this time we won't travel
                                 through its discriminator because we passed in
                                 _visited_composed_classes = (Animal,)
-            start_at (float): [optional]  # noqa: E501
+            start_at (int): [optional]  # noqa: E501
             description (str): [optional]  # noqa: E501
             channel (SendBirdChannelResponse): [optional]  # noqa: E501
-            end_at (float): [optional]  # noqa: E501
+            end_at (int): [optional]  # noqa: E501
         """
 
         _check_type = kwargs.pop('_check_type', True)
