+++ sendbird_platform_sdk/model/list_data_exports_by_message_channel_or_user_response_exported_data_inner.py
@@ -93,9 +93,9 @@ class ListDataExportsByMessageChannelOrUserResponseExportedDataInner(ModelNormal
             'format': (str,),  # noqa: E501
             'csv_delimiter': (str,),  # noqa: E501
             'timezone': (str,),  # noqa: E501
-            'created_at': (float,),  # noqa: E501
-            'start_ts': (float,),  # noqa: E501
-            'end_ts': (float,),  # noqa: E501
+            'created_at': (int,),  # noqa: E501
+            'start_ts': (int,),  # noqa: E501
+            'end_ts': (int,),  # noqa: E501
             'channel_urls': ([str],),  # noqa: E501
             'sender_ids': ([str],),  # noqa: E501
             'file': (ListDataExportsByMessageChannelOrUserResponseExportedDataInnerFile,),  # noqa: E501
@@ -168,9 +168,9 @@ class ListDataExportsByMessageChannelOrUserResponseExportedDataInner(ModelNormal
             format (str): [optional]  # noqa: E501
             csv_delimiter (str): [optional]  # noqa: E501
             timezone (str): [optional]  # noqa: E501
-            created_at (float): [optional]  # noqa: E501
-            start_ts (float): [optional]  # noqa: E501
-            end_ts (float): [optional]  # noqa: E501
+            created_at (int): [optional]  # noqa: E501
+            start_ts (int): [optional]  # noqa: E501
+            end_ts (int): [optional]  # noqa: E501
             channel_urls ([str]): [optional]  # noqa: E501
             sender_ids ([str]): [optional]  # noqa: E501
             file (ListDataExportsByMessageChannelOrUserResponseExportedDataInnerFile): [optional]  # noqa: E501
@@ -265,9 +265,9 @@ class ListDataExportsByMessageChannelOrUserResponseExportedDataInner(ModelNormal
             format (str): [optional]  # noqa: E501
             csv_delimiter (str): [optional]  # noqa: E501
             timezone (str): [optional]  # noqa: E501
-            created_at (float): [optional]  # noqa: E501
-            start_ts (float): [optional]  # noqa: E501
-            end_ts (float): [optional]  # noqa: E501
+            created_at (int): [optional]  # noqa: E501
+            start_ts (int): [optional]  # noqa: E501
+            end_ts (int): [optional]  # noqa: E501
             channel_urls ([str]): [optional]  # noqa: E501
             sender_ids ([str]): [optional]  # noqa: E501
             file (ListDataExportsByMessageChannelOrUserResponseExportedDataInnerFile): [optional]  # noqa: E501
