+++ sendbird_platform_sdk/model/oc_list_banned_users_response_banned_list_inner.py
@@ -89,8 +89,8 @@ class OcListBannedUsersResponseBannedListInner(ModelNormal):
         lazy_import()
         return {
             'user': (SendBirdUser,),  # noqa: E501
-            'start_at': (float,),  # noqa: E501
-            'end_at': (float,),  # noqa: E501
+            'start_at': (int,),  # noqa: E501
+            'end_at': (int,),  # noqa: E501
             'description': (str,),  # noqa: E501
         }
 
@@ -148,8 +148,8 @@ class OcListBannedUsersResponseBannedListInner(ModelNormal):
                                 through its discriminator because we passed in
                                 _visited_composed_classes = (Animal,)
             user (SendBirdUser): [optional]  # noqa: E501
-            start_at (float): [optional]  # noqa: E501
-            end_at (float): [optional]  # noqa: E501
+            start_at (int): [optional]  # noqa: E501
+            end_at (int): [optional]  # noqa: E501
             description (str): [optional]  # noqa: E501
         """
 
@@ -237,8 +237,8 @@ class OcListBannedUsersResponseBannedListInner(ModelNormal):
                                 through its discriminator because we passed in
                                 _visited_composed_classes = (Animal,)
             user (SendBirdUser): [optional]  # noqa: E501
-            start_at (float): [optional]  # noqa: E501
-            end_at (float): [optional]  # noqa: E501
+            start_at (int): [optional]  # noqa: E501
+            end_at (int): [optional]  # noqa: E501
             description (str): [optional]  # noqa: E501
         """
