+++ sendbird_platform_sdk/model/register_gdpr_request_response.py
@@ -88,7 +88,7 @@ class RegisterGdprRequestResponse(ModelNormal):
             'user_id': (str,),  # noqa: E501
             'user_ids': ([str],),  # noqa: E501
             'channel_delete_option': (str,),  # noqa: E501
-            'created_at': (float,),  # noqa: E501
+            'created_at': (int,),  # noqa: E501
         }
 
     @cached_property
@@ -153,7 +153,7 @@ class RegisterGdprRequestResponse(ModelNormal):
             user_id (str): [optional]  # noqa: E501
             user_ids ([str]): [optional]  # noqa: E501
             channel_delete_option (str): [optional]  # noqa: E501
-            created_at (float): [optional]  # noqa: E501
+            created_at (int): [optional]  # noqa: E501
         """
 
         _check_type = kwargs.pop('_check_type', True)
@@ -245,7 +245,7 @@ class RegisterGdprRequestResponse(ModelNormal):
             user_id (str): [optional]  # noqa: E501
             user_ids ([str]): [optional]  # noqa: E501
             channel_delete_option (str): [optional]  # noqa: E501
-            created_at (float): [optional]  # noqa: E501
+            created_at (int): [optional]  # noqa: E501
         """
 
         _check_type = kwargs.pop('_check_type', True)
